requirements.txt
fastapi
uvicorn
sqlalchemy
PyJWT
python-dotenv
psycopg2
bcrypt